More accurate statistics
Let client send statistics at 100hz, and dashboard to display it
Epoch only allows 1hz, I replaced it with uPlot, little less pretty but more functional

+++ alvr/client/android/ALVR-common/packet_types.h
@@ -234,7 +234,7 @@ struct TimeSync {
 	uint64_t fecFailureInSecond;
 	uint64_t fecFailureTotal;
 
-	uint32_t fps;
+	float fps;
 };
 struct VideoFrame {
 	uint32_t type; // ALVR_PACKET_TYPE_VIDEO_FRAME
